[FIX] server/command_ai => look function was a bit off
had a size_t instead of a ssize_t so index was overflowing for negative
values and never showing tiles

also modified code to suit epitech coding style

+++ server/src/server/summons/command_ai/look_func.c
@@ -15,8 +15,8 @@
 static int look_west(server_t *server, client_node_t *client)
 {
     dprint_tile(server->map, client->stats.pos.x, client->stats.pos.y, client);
-    for (size_t i = 1; i <= client->stats.level; i += 1) {
-        for (size_t j = -i; j <= (i + 1); j += 1) {
+    for (ssize_t i = 1; i <= (ssize_t)client->stats.level; i += 1) {
+        for (ssize_t j = -i; j <= (i + 1); j += 1) {
             dprintf(client->cfd, ",");
             dprint_tile(server->map, client->stats.pos.x - j,
                         client->stats.pos.y - i, client);
@@ -28,8 +28,8 @@ static int look_west(server_t *server, client_node_t *client)
 static int look_east(server_t *server, client_node_t *client)
 {
     dprint_tile(server->map, client->stats.pos.x, client->stats.pos.y, client);
-    for (size_t i = 1; i <= client->stats.level; i += 1) {
-        for (size_t j = -i; j <= (i + 1); j += 1) {
+    for (ssize_t i = 1; i <= (ssize_t)client->stats.level; i += 1) {
+        for (ssize_t j = -i; j <= (i + 1); j += 1) {
             dprintf(client->cfd, ",");
             dprint_tile(server->map, client->stats.pos.x + j,
                         client->stats.pos.y + i, client);
@@ -41,8 +41,8 @@ static int look_east(server_t *server, client_node_t *client)
 static int look_south(server_t *server, client_node_t *client)
 {
     dprint_tile(server->map, client->stats.pos.x, client->stats.pos.y, client);
-    for (size_t i = 1; i <= client->stats.level; i += 1) {
-        for (size_t j = -i; j <= (i + 1); j += 1) {
+    for (ssize_t i = 1; i <= (ssize_t)client->stats.level; i += 1) {
+        for (ssize_t j = -i; j <= (i + 1); j += 1) {
             dprintf(client->cfd, ",");
             dprint_tile(server->map, client->stats.pos.x - i,
                         client->stats.pos.y + j, client);
@@ -54,8 +54,8 @@ static int look_south(server_t *server, client_node_t *client)
 static int look_north(server_t *server, client_node_t *client)
 {
     dprint_tile(server->map, client->stats.pos.x, client->stats.pos.y, client);
-    for (size_t i = 1; i <= client->stats.level; i += 1) {
-        for (size_t j = -i; j <= (i + 1); j += 1) {
+    for (ssize_t i = 1; i <= (ssize_t)client->stats.level; i += 1) {
+        for (ssize_t j = -i; j <= (i + 1); j += 1) {
             dprintf(client->cfd, ",");
             dprint_tile(server->map, client->stats.pos.x + i,
                         client->stats.pos.y - j, client);
